Refactor creation of IC3Adapter
